requirements.txt
apache-airflow==2.9.2
Flask==2.2.5
pandas==2.2.2
SQLAlchemy==1.4.52
requests==2.32.3
psycopg2-binary==2.9.9